feat: introduce storage configuration on reworked internal

+++ src/testRecipes/java/support/Assertions.java
@@ -5,7 +5,6 @@ import org.apache.commons.io.FileUtils;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.function.Consumer;
 
@@ -22,6 +21,15 @@ public class Assertions {
     public static void assertBaselineUsed(Consumer<String> assertion, String baselinePath) {
         var dummyData = "my data";
         var baselineFile = new File(baselinePath);
+
+        if (baselineFile.exists()) {
+            throw new AssertionError(
+                ("baseline %s exists prior to test execution. Check and remove manually." +
+                    "\nThis would lead to unexpected positive results")
+                    .formatted(baselineFile.getAbsolutePath())
+            );
+        }
+
         try {
             FileUtils.writeStringToFile(baselineFile, dummyData, StandardCharsets.UTF_8);
             assertion.accept(dummyData);
